[ver 1.01.88] Refactoring TktkDX12WrappingLib Code

+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/Buffer/DepthStencil/DepthStencilBuffer.h
@@ -17,19 +17,19 @@ namespace tktk
 	public:
 
 		// 「DepthStencilBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int create(ID3D12Device* device, const DepthStencilBufferInitParam& initParam);
+		size_t create(ID3D12Device* device, const DepthStencilBufferInitParam& initParam);
 
 		// 指定の深度ステンシルバッファを削除する
 		// ※引数のハンドルに対応するリソースが無かったら何もしない
-		void erase(unsigned int handle);
+		void erase(size_t handle);
 
 		// 指定の深度ステンシルバッファのリソースバリアを深度書き込み状態に変更する
 		// ※シェーダーリソースとして使用しない設定の場合読んでも何も起きない
-		void beginWrite(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void beginWrite(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定の深度ステンシルバッファのリソースバリアをシェーダー使用状態に変更する
 		// ※シェーダーリソースとして使用しない設定の場合読んでも何も起きない
-		void endWrite(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void endWrite(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 全ての深度ステンシルバッファのリソースバリアを深度書き込み状態に変更する
 		// ※シェーダーリソースとして使用しない設定の場合読んでも何も起きない
@@ -40,17 +40,17 @@ namespace tktk
 		void allEndWrite(ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定の深度ステンシルバッファを使用して、引数のディスクリプタハンドルに深度ステンシルビューを作る
-		void createDsv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
+		void createDsv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
 
 		// 指定の深度ステンシルバッファを使用して、引数のディスクリプタハンドルにシェーダーリソースビューを作る
 		// ※シェーダーリソースとして使用しない設定の場合、エラーを吐きます
-		void createSrv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
+		void createSrv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
 
 		// 指定の深度ステンシルバッファ画像の大きさを取得する（ピクセル）
-		const tktkMath::Vector2& getDepthStencilSizePx(unsigned int handle) const;
+		const tktkMath::Vector2& getDepthStencilSizePx(size_t handle) const;
 
 		// 指定の深度ステンシルバッファのポインタを取得する
-		ID3D12Resource* getBufferPtr(unsigned int handle) const;
+		ID3D12Resource* getBufferPtr(size_t handle) const;
 
 	private:
