+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/Buffer/Index/IndexBuffer.h
@@ -17,17 +17,17 @@ namespace tktk
 	public:
 
 		// 「IndexBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int create(ID3D12Device* device, const std::vector<unsigned short>& indexDataArray);
+		size_t create(ID3D12Device* device, const std::vector<unsigned short>& indexDataArray);
 
 		// 指定のインデックスバッファを削除する
 		// ※引数のハンドルに対応するリソースが無かったら何もしない
-		void erase(unsigned int handle);
+		void erase(size_t handle);
 
 		// コマンドリストに指定のインデックスバッファを登録する
-		void set(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void set(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定のインデックスバッファのポインタを取得する
-		ID3D12Resource* getBufferPtr(unsigned int handle) const;
+		ID3D12Resource* getBufferPtr(size_t handle) const;
 
 	private:
