+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/Buffer/RenderTarget/RenderTargetBuffer.h
@@ -17,38 +17,38 @@ namespace tktk
 	public:
 
 		// ゼロから「RenderTargetBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int create(ID3D12Device* device, const tktkMath::Vector2& renderTargetSize, const tktkMath::Color& clearColor);
+		size_t create(ID3D12Device* device, const tktkMath::Vector2& renderTargetSize, const tktkMath::Color& clearColor);
 
 		// スワップチェインから「RenderTargetBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int create(IDXGISwapChain1* swapChain, unsigned int backBufferIndex);
+		size_t create(IDXGISwapChain1* swapChain, unsigned int backBufferIndex);
 
 		// 指定のレンダーターゲットバッファを削除する
 		// ※引数のハンドルに対応するリソースが無かったら何もしない
-		void erase(unsigned int handle);
+		void erase(size_t handle);
 
 		// 指定のレンダーターゲットバッファのリソースバリアをレンダーターゲット状態に変更する
-		void beginWriteBasicRtBuffer(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void beginWriteBasicRtBuffer(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定のレンダーターゲットバッファのリソースバリアをシェーダー使用状態に変更する
-		void endWriteBasicRtBuffer(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void endWriteBasicRtBuffer(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定のレンダーターゲットバッファのリソースバリアをレンダーターゲット状態に変更する
-		void beginWriteBackBuffer(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void beginWriteBackBuffer(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定のレンダーターゲットバッファのリソースバリアをプリセット状態に変更する
-		void endWriteBackBuffer(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void endWriteBackBuffer(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 		// 指定のレンダーターゲットバッファを使用して、引数のディスクリプタハンドルにレンダーターゲットビューを作る
-		void createRtv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
+		void createRtv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
 
 		// 指定のレンダーターゲットバッファを使用して、引数のディスクリプタハンドルにシェーダーリソースビューを作る
-		void createSrv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
+		void createSrv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
 
 		// 指定のレンダーターゲットバッファ画像の大きさを取得す（ピクセル）
-		const tktkMath::Vector2& getRenderTargetSizePx(unsigned int handle) const;
+		const tktkMath::Vector2& getRenderTargetSizePx(size_t handle) const;
 
 		// 指定のレンダーターゲットバッファのポインタを取得する
-		ID3D12Resource* getBufferPtr(unsigned int handle) const;
+		ID3D12Resource* getBufferPtr(size_t handle) const;
 
 	private:
