+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/Buffer/Texture/TextureBuffer.h
@@ -18,29 +18,29 @@ namespace tktk
 	public:
 		
 		// コマンドリストを使わずに「TextureBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int cpuPriorityCreate(ID3D12Device* device, const TexBufFormatParam& formatParam, const TexBuffData& dataParam);
+		size_t cpuPriorityCreate(ID3D12Device* device, const TexBufFormatParam& formatParam, const TexBuffData& dataParam);
 		
 		// コマンドリストを使って「TextureBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int gpuPriorityCreate(ID3D12Device* device, ID3D12GraphicsCommandList* commandList, const TexBufFormatParam& formatParam, const TexBuffData& dataParam);
+		size_t gpuPriorityCreate(ID3D12Device* device, ID3D12GraphicsCommandList* commandList, const TexBufFormatParam& formatParam, const TexBuffData& dataParam);
 
 		// 引数のファイルから画像情報をロードし、コマンドリストを使わずに「TextureBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int cpuPriorityLoad(ID3D12Device* device, const std::string& texDataPath);
+		size_t cpuPriorityLoad(ID3D12Device* device, const std::string& texDataPath);
 
 		// 引数のファイルから画像情報をロードし、コマンドリストを使って「TextureBufferData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int gpuPriorityLoad(ID3D12Device* device, ID3D12GraphicsCommandList* commandList, const std::string& texDataPath);
+		size_t gpuPriorityLoad(ID3D12Device* device, ID3D12GraphicsCommandList* commandList, const std::string& texDataPath);
 	
 		// 指定のテクスチャバッファを削除する
 		// ※引数のハンドルに対応するリソースが無かったら何もしない
-		void erase(unsigned int handle);
+		void erase(size_t handle);
 
 		// 指定のテクスチャバッファを使用して、引数のディスクリプタハンドルにシェーダーリソースビューを作る
-		void createSrv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
+		void createSrv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const;
 		
 		// 指定のテクスチャバッファ画像の大きさを取得する（ピクセル）
-		const tktkMath::Vector3& getTextureSizePx(unsigned int handle) const;
+		const tktkMath::Vector3& getTextureSizePx(size_t handle) const;
 
 		// 指定のテクスチャバッファのポインタを取得する
-		ID3D12Resource* getBufferPtr(unsigned int handle) const;
+		ID3D12Resource* getBufferPtr(size_t handle) const;
 
 	private:
