+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/DescriptorHeap/Basic/BasicDescriptorHeapData.h
@@ -32,8 +32,8 @@ namespace tktk
 	public:
 
 		// それぞれのディスクリプタテーブルが持っているディスクリプタの数の配列
-		std::vector<unsigned int>	m_descriptorTableSizeArray;
-		ID3D12DescriptorHeap*		m_descriptorHeap			{ nullptr };
+		std::vector<size_t>		m_descriptorTableSizeArray;
+		ID3D12DescriptorHeap*	m_descriptorHeap			{ nullptr };
 	};
 }
 #endif // !BASIC_DESCRIPTOR_HEAP_DATA_H_
