+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/DescriptorHeap/Basic/BasicDescriptorHeapInitParam.h
@@ -10,7 +10,7 @@ namespace tktk
 	struct BasicDescriptorParam
 	{
 		BufferType			type	{};
-		unsigned int		handle	{};
+		size_t				handle	{};
 	};
 
 	// このディスクリプタヒープで扱うバッファーの種類
