+++ TktkDirectX12GameLib/TktkDX12WrappingLib/inc/TktkDX12Wrapper/Resource/ScissorRect/ScissorRect.h
@@ -17,14 +17,14 @@ namespace tktk
 	public:
 
 		// 「ScissorRectData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int create(const std::vector<ScissorRectInitParam>& initParamArray);
+		size_t create(const std::vector<ScissorRectInitParam>& initParamArray);
 
 		// 指定のシザー矩形を削除する
 		// ※引数のハンドルに対応するリソースが無かったら何もしない
-		void erase(unsigned int handle);
+		void erase(size_t handle);
 
 		// 指定のシザー矩形をコマンドリストに登録する
-		void set(unsigned int handle, ID3D12GraphicsCommandList* commandList) const;
+		void set(size_t handle, ID3D12GraphicsCommandList* commandList) const;
 
 	private:
