+++ TktkDirectX12GameLib/TktkDX12WrappingLib/src/TktkDX12Wrapper/Resource/Buffer/DepthStencil/DepthStencilBuffer.cpp
@@ -7,22 +7,22 @@ namespace tktk
 	{
 	}
 
-	unsigned int DepthStencilBuffer::create(ID3D12Device* device, const DepthStencilBufferInitParam& initParam)
+	size_t DepthStencilBuffer::create(ID3D12Device* device, const DepthStencilBufferInitParam& initParam)
 	{
 		return m_depthStencilBufferDataArray.create(device, initParam);
 	}
 
-	void DepthStencilBuffer::erase(unsigned int handle)
+	void DepthStencilBuffer::erase(size_t handle)
 	{
 		m_depthStencilBufferDataArray.erase(handle);
 	}
 
-	void DepthStencilBuffer::beginWrite(unsigned int handle, ID3D12GraphicsCommandList* commandList) const
+	void DepthStencilBuffer::beginWrite(size_t handle, ID3D12GraphicsCommandList* commandList) const
 	{
 		m_depthStencilBufferDataArray.getMatchHandlePtr(handle)->beginWrite(commandList);
 	}
 
-	void DepthStencilBuffer::endWrite(unsigned int handle, ID3D12GraphicsCommandList* commandList) const
+	void DepthStencilBuffer::endWrite(size_t handle, ID3D12GraphicsCommandList* commandList) const
 	{
 		m_depthStencilBufferDataArray.getMatchHandlePtr(handle)->endWrite(commandList);
 	}
@@ -43,22 +43,22 @@ namespace tktk
 		}
 	}
 
-	void DepthStencilBuffer::createDsv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const
+	void DepthStencilBuffer::createDsv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const
 	{
 		m_depthStencilBufferDataArray.getMatchHandlePtr(handle)->createDsv(device, heapHandle);
 	}
 
-	void DepthStencilBuffer::createSrv(unsigned int handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const
+	void DepthStencilBuffer::createSrv(size_t handle, ID3D12Device* device, D3D12_CPU_DESCRIPTOR_HANDLE heapHandle) const
 	{
 		m_depthStencilBufferDataArray.getMatchHandlePtr(handle)->createSrv(device, heapHandle);
 	}
 
-	const tktkMath::Vector2& DepthStencilBuffer::getDepthStencilSizePx(unsigned int handle) const
+	const tktkMath::Vector2& DepthStencilBuffer::getDepthStencilSizePx(size_t handle) const
 	{
 		return m_depthStencilBufferDataArray.getMatchHandlePtr(handle)->getDepthStencilSizePx();
 	}
 
-	ID3D12Resource* DepthStencilBuffer::getBufferPtr(unsigned int handle) const
+	ID3D12Resource* DepthStencilBuffer::getBufferPtr(size_t handle) const
 	{
 		return m_depthStencilBufferDataArray.getMatchHandlePtr(handle)->getBufferPtr();
 	}
