+++ TktkDirectX12GameLib/TktkDX12WrappingLib/src/TktkDX12Wrapper/Resource/Buffer/Index/IndexBuffer.cpp
@@ -7,22 +7,22 @@ namespace tktk
 	{
 	}
 
-	unsigned int IndexBuffer::create(ID3D12Device* device, const std::vector<unsigned short>& indexDataArray)
+	size_t IndexBuffer::create(ID3D12Device* device, const std::vector<unsigned short>& indexDataArray)
 	{
 		return m_indexBufferDataArray.create(device, indexDataArray);
 	}
 
-	void IndexBuffer::erase(unsigned int handle)
+	void IndexBuffer::erase(size_t handle)
 	{
 		m_indexBufferDataArray.erase(handle);
 	}
 
-	void IndexBuffer::set(unsigned int handle, ID3D12GraphicsCommandList* commandList) const
+	void IndexBuffer::set(size_t handle, ID3D12GraphicsCommandList* commandList) const
 	{
 		m_indexBufferDataArray.getMatchHandlePtr(handle)->set(commandList);
 	}
 
-	ID3D12Resource* IndexBuffer::getBufferPtr(unsigned int handle) const
+	ID3D12Resource* IndexBuffer::getBufferPtr(size_t handle) const
 	{
 		return m_indexBufferDataArray.getMatchHandlePtr(handle)->getBufferPtr();
 	}
